refactor(auth): extract credential and integration auth subsystem module
